4.3

+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -79,7 +79,7 @@ public class StudentController {
     }
 
     @GetMapping(value = "/avg-age")
-    public ResponseEntity<Integer> getAverageAgeOfStudents() {
+    public ResponseEntity<Double> getAverageAgeOfStudents() {
         return ResponseEntity.ok(studentService.getAverageAgeOfStudents());
     }
